Refactor main and main test
Reduce individual file size. Make it more manageable
