+++ src/pitloom/cli/options.py
@@ -1,0 +1,291 @@
+# SPDX-FileContributor: Arthit Suriyawongkul
+# SPDX-FileCopyrightText: 2026-present Arthit Suriyawongkul
+# SPDX-FileType: SOURCE
+# SPDX-License-Identifier: Apache-2.0
+
+"""Command-line interface for Pitloom's SBOM generator."""
+
+from __future__ import annotations
+
+import argparse
+import sys
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Any
+
+from pitloom.core.config import PitloomConfig
+from pitloom.core.creation import (
+    CreationMetadata,
+    Creator,
+    Tool,
+)
+from pitloom.core.project import ProjectMetadata
+from pitloom.extract.project import read_project
+
+_SPDX3_JSON_EXT = ".spdx3.json"
+_PROJECT_PYPROJECT_SOURCE = "pyproject.toml"
+_PROJECT_SETUP_CFG_SOURCE = "setup.cfg"
+_PROJECT_SETUP_PY_SOURCE = "setup.py"
+
+
+@dataclass(frozen=True)
+class _ResolvedValue:
+    """A resolved option value paired with its source label."""
+
+    value: str | None
+    source: str
+
+
+@dataclass(frozen=True)
+class _ResolvedCreators:
+    """Resolved creator list paired with its source label."""
+
+    value: list[Creator]
+    source: str
+
+
+@dataclass(frozen=True)
+class _ResolvedTools:
+    """Resolved tool list paired with its source label."""
+
+    value: list[Tool] | None
+    source: str
+
+
+@dataclass(frozen=True)
+class _ResolvedCreationMetadata:
+    """Resolved creation metadata values and their source labels."""
+
+    creators: _ResolvedCreators
+    tools: _ResolvedTools
+    creation_datetime: _ResolvedValue
+    creation_comment: _ResolvedValue
+
+    def to_creation_metadata(self) -> CreationMetadata:
+        """Convert resolved values to :class:`CreationMetadata`."""
+        return CreationMetadata(
+            creators=self.creators.value,
+            tools=self.tools.value,
+            creation_datetime=self.creation_datetime.value,
+            creation_comment=self.creation_comment.value,
+        )
+
+
+def _resolve_project_paths(args: argparse.Namespace) -> tuple[Path | None, Path | None]:
+    """Resolve and validate project directory or sdist archive path."""
+    project_dir = args.project_dir.resolve()
+    if not project_dir.exists():
+        print(f"ERROR: project directory not found: {project_dir}", file=sys.stderr)
+        return None, None
+
+    if project_dir.is_file():
+        return project_dir, project_dir
+
+    for candidate in ("pyproject.toml", "setup.cfg", "setup.py"):
+        config_path = project_dir / candidate
+        if config_path.exists():
+            return project_dir, config_path
+
+    print(
+        f"ERROR: no project configuration found in {project_dir}. "
+        "Expected pyproject.toml, setup.cfg, or setup.py.",
+        file=sys.stderr,
+    )
+    return None, None
+
+
+def _resolve_creation_field(
+    cli_value: str | None,
+    config_value: str | None,
+    default_value: str | None,
+) -> _ResolvedValue:
+    """Resolve a creation field with precedence CLI > pyproject > default."""
+    if cli_value is not None:
+        return _ResolvedValue(value=cli_value, source="command-line")
+    if config_value is not None:
+        return _ResolvedValue(value=config_value, source=_PROJECT_PYPROJECT_SOURCE)
+    return _ResolvedValue(value=default_value, source="default")
+
+
+def _resolve_creators(
+    args: argparse.Namespace,
+    config_creators: list[Creator],
+) -> _ResolvedCreators:
+    """Resolve the creator list with precedence CLI > pyproject > default."""
+    cli_creators: list[Creator] | None = args.creators
+    if cli_creators:
+        return _ResolvedCreators(value=cli_creators, source="command-line")
+    if config_creators:
+        return _ResolvedCreators(
+            value=config_creators, source=_PROJECT_PYPROJECT_SOURCE
+        )
+    return _ResolvedCreators(value=[], source="default")
+
+
+def _resolve_tools(
+    args: argparse.Namespace,
+    config_tools: list[Tool] | None,
+) -> _ResolvedTools:
+    """Resolve the tool list."""
+    if args.no_creation_tool:
+        return _ResolvedTools(value=[], source="command-line")
+    cli_tools: list[str] | None = args.creation_tools
+    if cli_tools:
+        return _ResolvedTools(
+            value=[Tool(name=name) for name in cli_tools],
+            source="command-line",
+        )
+    if config_tools is not None:
+        return _ResolvedTools(value=config_tools, source=_PROJECT_PYPROJECT_SOURCE)
+    return _ResolvedTools(value=None, source="default")
+
+
+def _resolve_creation_metadata(
+    args: argparse.Namespace,
+    pitloom_config: Any,
+) -> _ResolvedCreationMetadata:
+    """Resolve creation metadata."""
+    default_creation = CreationMetadata()
+    return _ResolvedCreationMetadata(
+        creators=_resolve_creators(args, pitloom_config.creators),
+        tools=_resolve_tools(args, pitloom_config.tools),
+        creation_datetime=_resolve_creation_field(
+            args.creation_datetime,
+            pitloom_config.creation_datetime,
+            default_creation.creation_datetime,
+        ),
+        creation_comment=_resolve_creation_field(
+            args.creation_comment,
+            pitloom_config.creation_comment,
+            "Generated via Pitloom CLI",
+        ),
+    )
+
+
+def _resolve_generate_mode_settings(
+    args: argparse.Namespace,
+) -> tuple[CreationMetadata, bool | None, bool | None]:
+    """Resolve settings for ``loom generate`` using project config when available."""
+    pitloom_config = PitloomConfig()
+    target_value = str(args.target).strip()
+    if target_value:
+        target_path = Path(target_value)
+        if target_path.exists():
+            try:
+                _, pitloom_config, _ = read_project(target_path)
+            except FileNotFoundError:
+                pitloom_config = PitloomConfig()
+
+    creation = _resolve_creation_metadata(args, pitloom_config)
+    effective_pretty = pitloom_config.pretty if args.pretty is None else args.pretty
+    effective_describe_relationship = (
+        pitloom_config.describe_relationship
+        if args.describe_relationship is None
+        else args.describe_relationship
+    )
+    return (
+        creation.to_creation_metadata(),
+        effective_pretty,
+        effective_describe_relationship,
+    )
+
+
+def _load_pitloom_tool_section(config_path: Path | None) -> dict[str, Any]:
+    """Load ``[tool.pitloom]`` keys for verbose source reporting."""
+    if config_path is None or config_path.name != "pyproject.toml":
+        return {}
+
+    # pylint: disable=import-outside-toplevel
+    if sys.version_info >= (3, 11):
+        import tomllib
+    else:
+        import tomli as tomllib
+
+    try:
+        raw_toml = tomllib.loads(config_path.read_text(encoding="utf-8"))
+        tool_section = raw_toml.get("tool")
+        if not isinstance(tool_section, dict):
+            return {}
+        pitloom_tool = tool_section.get("pitloom")
+        if not isinstance(pitloom_tool, dict):
+            return {}
+        return {str(key): value for key, value in pitloom_tool.items()}
+    except Exception:  # pylint: disable=broad-exception-caught
+        return {}
+
+
+def _resolve_output_source(
+    args: argparse.Namespace, pitloom_config: Any, config_path: Path | None
+) -> str:
+    if args.output is not None:
+        return "command-line"
+    if pitloom_config.sbom_basename:
+        return config_path.name if config_path else _PROJECT_PYPROJECT_SOURCE
+    return "default"
+
+
+def _resolve_pretty(
+    args: argparse.Namespace,
+    pitloom_config: Any,
+    pitloom_tool: dict[str, Any],
+    config_source: str = _PROJECT_PYPROJECT_SOURCE,
+) -> tuple[bool, str]:
+    value = pitloom_config.pretty if args.pretty is None else args.pretty
+    if args.pretty is not None:
+        return value, "command-line"
+    if "pretty" in pitloom_tool:
+        return value, config_source
+    return value, "default"
+
+
+def _resolve_describe_relationship(
+    args: argparse.Namespace,
+    pitloom_config: Any,
+    pitloom_tool: dict[str, Any],
+    config_source: str = _PROJECT_PYPROJECT_SOURCE,
+) -> tuple[bool, str]:
+    value = bool(
+        pitloom_config.describe_relationship
+        if args.describe_relationship is None
+        else args.describe_relationship
+    )
+    if args.describe_relationship is not None:
+        return value, "command-line"
+    if (
+        "describe_relationship" in pitloom_tool
+        or "describe-relationship" in pitloom_tool
+    ):
+        return value, config_source
+    return value, "default"
+
+
+def _quote_optional(value: str | None) -> str:
+    if value is None:
+        return "None"
+    return f"'{value}'"
+
+
+def _resolve_output_path(
+    explicit: Path | None, metadata: ProjectMetadata, pitloom_config: Any
+) -> Path:
+    if explicit is not None:
+        return explicit
+    if pitloom_config.sbom_basename:
+        return Path(f"{pitloom_config.sbom_basename}{_SPDX3_JSON_EXT}")
+    parts = [metadata.name] if metadata.name else ["sbom"]
+    if metadata.version:
+        parts.append(metadata.version)
+    return Path("-".join(parts) + _SPDX3_JSON_EXT)
+
+
+def _resolve_model_output_path(explicit: Path | None, model_path: Path) -> Path:
+    if explicit is not None:
+        return explicit
+    return Path.cwd() / (model_path.name + _SPDX3_JSON_EXT)
+
+
+def _resolve_hf_output_path(explicit: Path | None, model_id: str) -> Path:
+    if explicit is not None:
+        return explicit
+    stem = model_id.split("/")[-1]
+    return Path.cwd() / (stem + _SPDX3_JSON_EXT)
